added cheking if deleting of student is available

+++ TestFactory/Controllers/Api/StudentsController.cs
@@ -11,6 +11,7 @@ using TestFactory.Filters;
 using TestFactory.MVC.ViewModels;
 using System;
 using TestFactory.Controllers;
+using TestFactory.Business.Components;
 
 namespace TestFactory.Controllers.Api
 {
@@ -19,11 +20,12 @@ namespace TestFactory.Controllers.Api
     {
         private readonly StudentManager studentManager;
         private readonly GroupManager groupManager;
-
+        private UserContext user;
         public StudentsController(StudentManager studentManager, GroupManager groupManager)
         {
             this.studentManager = studentManager;
             this.groupManager = groupManager;
+            this.user = new UserContext();
         }
 
         [HttpGet]
@@ -74,6 +76,8 @@ namespace TestFactory.Controllers.Api
         [Route("delete")]
         public IHttpActionResult Delete(StudentViewModel student)
         {
+            if (!groupManager.HasAccessToGroup(student.GroupId, user.User.Id))
+                return Ok();
             var model = Mapper.Map<Student>(student);
             studentManager.Delete(model.Id);
             return Ok();
